requirements.txt
openai
groq
ollama
requests
python-dotenv

